bull and cow game
This is an guessing number game. The user has to guess the number correctly., The  player writes a 4-digit number that they think might be the correct answer. Then, the computer grades that answer using bulls and cows as a notation. If a digit in the given answer matches a digit and its position in the code, it's called a "bull." If a given digit appears in the code but it position doesn't match, then it's called a "cow." The computer reveals how many bulls and cows there are. The information is general; in other words, it isn't bound to any particular digit.